+++ data/beaconrestapi/src/main/java/tech/pegasys/teku/beaconrestapi/BeaconRestApi.java
@@ -435,11 +435,7 @@ public class BeaconRestApi {
     app.get(
         GetEvents.ROUTE,
         new GetEvents(
-            dataProvider,
-            jsonProvider,
-            eventChannels,
-            asyncRunner,
-            configuration.getMaxPendingEvents()));
+            dataProvider, eventChannels, asyncRunner, configuration.getMaxPendingEvents()));
   }
 
   public void stop() {